requirements.txt
gradio
pandas
snowflake-connector-python
snowflake-sqlalchemy
langchain
langchain_community
openai
